openSense layer standardized
