+++ example/layers.js
@@ -92,7 +92,7 @@ var bounds = new L.LatLngBounds(
     var AQICNLayer = L.layerGroup.aqicnLayer();
     var OpenAqLayer = L.layerGroup.layerCode("openaq");
     var LuftdatenLayer = L.layerGroup.layerCode("luftdaten");
-    var OpenSenseLayer = L.layerGroup.openSenseLayer();
+    var OpenSenseLayer = L.layerGroup.layerCode("opensense");
     var city = L.OWM.current({ intervall: 15, minZoom: 3 });
     var windrose = L.OWM.current({
         intervall: 15,
